Hello,

I sat down and tried to get a bit more performance out of the script.
Baseline, vertex colors only: 23,8 fps
I made the following changes, one per commit:
- min/max clamping of weights per lambda: 24,08 fps (It's really hard to read imo, if this can be formatted in a better way, I'd be glad to improve it!)
- vertex_color as 2D data structure -> easier to read and 25,21 fps (with only this change)
- added global object for original edge length -> 25,8 fps (with only this change)
- added index mapping: from the loop-based index to a vertex-based index, this also simplifies the setting of the vertex colors -> 28,37fps (with only this change)

I now get 31,32fps with all those changes together.

I'll mark this as draft for now, since I want to look into it a bit more, but your input would be appreciated!