@ScottishCyclops 2019-09-08
Thank you !
I'm about to read your code. That sounds promising.