@VincentKueszter 2019-09-08
> > It's really hard to read imo, if this can be formatted in a better way, I'd be glad to improve it!
> 
> I've always found the python way of writing inline ifs (ternary) very readable, as in you can read the line of code in English and it makes sense.
> 
> No problem with readability in my opinion.
> 
> I'm glad to hear this increases performance, because it clearly increases the overall readability too 👍

If I could, I would format it like this. It's a bit easier to see what happens in each case at a glance, in my opinion.
```
lower if value < lower else 
upper if value > upper else 
value
```