@VincentKueszter 2019-09-08
> Here is what I propose:
> This is an extra boost of real-time performance, for the ones that really need it. So why not use the same principle as the "cache" functionality found in many places in Blender ?
> You would manually choose to "cache" a mesh for TensionMap, enabling this performance gain.
> But since you manually chose to enable cache, then it's okay to have you update the cache when you change the mesh.
> That way, most users have the current behaviour of auto updating every frame, but don't have the hassle of updating the mesh in TensionMap manually everytime.
> And the hardcode real-time users can choose the functionality, knowing they will need to manually update.

Caching is a really fitting name for it! Sounds good.

> If you don't want to work on the UI part of this, let me know, I'll gladly do it.
> I'll take your current work here, turn it into a function, and link it to the UI.

I'll try it myself next week, the UI part didn't seem too complicated.