@ScottishCyclops 2019-09-08
> I'll try it myself next week, the UI part didn't seem too complicated.

Alright.
I'm picturing a button that says "Cache Geometry Data" with a hover text of something like "Improve realtime performance by pre-calculating some geometry data. An update to the mesh will require a manual cache update".

When the cache is enabled, the button is replaced by two buttons (in the same line maybe) that say "Disable cache" and "Update Cache".

The update of the Cache would need to stop being linked with `tm_update`, because it would have it's own operator.

Have fun coding :)