@VincentKueszter 2019-09-15
Okay, I re-implemented it as a cache.
I used dataclass instead of named tuple because it was a tiny bit faster.
I also updated the GUI and GUI code a bit, the row/col variable names were swapped and I removed some duplicated strings.
There should be no further performance improvements, even though I added a few more variables to avoid dot-chains.
Hope I didn't miss anything!