@VincentKueszter 2019-09-21
I added some more performance improvements by using bMesh:
7,9% more fps with vertex colors only and 10,2% with vertex groups only on my machine.