@VincentKueszter 2019-11-15
Hey, did you have time to look over it yet?